Moved SIMPL conversion functions to cpp files for their parameter
