+++ src/simplnx/Parameters/CreateColorMapParameter.cpp
@@ -62,4 +62,35 @@ Result<> CreateColorMapParameter::validate(const std::any& value) const
   [[maybe_unused]] const auto& json = GetAnyRef<ValueType>(value);
   return {};
 }
+
+namespace SIMPLConversion
+{
+namespace
+{
+constexpr StringLiteral k_SelectedPresetKey = "SelectedPresetName";
+constexpr StringLiteral k_PresetNameKey = "Name";
+constexpr StringLiteral k_RGBPointsKey = "RGBPoints";
+} // namespace
+
+Result<CreateColorMapFilterParameterConverter::ValueType> CreateColorMapFilterParameterConverter::convert(const nlohmann::json& json1, const nlohmann::json& json2)
+{
+  if(!json1.is_string())
+  {
+    return MakeErrorResult<ValueType>(-2, fmt::format("CreateColorMapFilterParameterConverter json1 '{}' is not a string", json2.dump()));
+  }
+  if(!json2.is_array())
+  {
+    return MakeErrorResult<ValueType>(-1, fmt::format("CreateColorMapFilterParameterConverter json2 '{}' is not an array", json1.dump()));
+  }
+
+  auto presetName = json1.get<std::string>();
+  auto rgbPoints = json2.get<std::vector<float64>>();
+
+  nlohmann::json value = nlohmann::json::object();
+  value[k_PresetNameKey] = presetName;
+  value[k_RGBPointsKey] = rgbPoints;
+
+  return {std::move(value)};
+}
+} // namespace SIMPLConversion
 } // namespace nx::core
