+++ src/simplnx/Parameters/NumericTypeParameter.cpp
@@ -1,5 +1,8 @@
 #include "NumericTypeParameter.hpp"
 
+#include "simplnx/Common/TypesUtility.hpp"
+#include "simplnx/Parameters/DataTypeParameter.hpp"
+
 #include <fmt/core.h>
 #include <nlohmann/json.hpp>
 
@@ -110,4 +113,50 @@ Result<> NumericTypeParameter::validate(const std::any& value) const
   [[maybe_unused]] auto castValue = std::any_cast<ValueType>(value);
   return {};
 }
+
+namespace SIMPLConversion
+{
+Result<NumericTypeParameterConverter::ValueType> NumericTypeParameterConverter::convert(const nlohmann::json& json)
+{
+  if(!json.is_number_integer())
+  {
+    return MakeErrorResult<ValueType>(-1, fmt::format("NumericTypeParameter json '{}' is not an integer", json.dump()));
+  }
+
+  auto value = json.get<int32>();
+
+  if(value < 0 || value > 10)
+  {
+    return MakeErrorResult<ValueType>(-2, fmt::format("Invalid NumericTypeParameter value '{}'", value));
+  }
+
+  // Convert size_t to uint64
+  if(value == 10)
+  {
+    return {NumericType::uint64};
+  }
+
+  return {static_cast<NumericType>(value)};
+}
+
+Result<ScalarTypeParameterToNumericTypeConverter::ValueType> ScalarTypeParameterToNumericTypeConverter::convert(const nlohmann::json& json)
+{
+  auto dataTypeResult = SIMPLConversion::ScalarTypeParameterConverter::convert(json);
+  if(dataTypeResult.invalid())
+  {
+    return ConvertInvalidResult<ValueType>(std::move(dataTypeResult));
+  }
+
+  DataType dataType = dataTypeResult.value();
+
+  std::optional<NumericType> numericType = ConvertDataTypeToNumericType(dataType);
+
+  if(!numericType.has_value())
+  {
+    return MakeErrorResult<ValueType>(-1, fmt::format("ScalarTypeParameter unable to convert DataType '{}' to NumericType", DataTypeToString(dataType)));
+  }
+
+  return {*numericType};
+}
+} // namespace SIMPLConversion
 } // namespace nx::core
