+++ src/simplnx/Parameters/ReadCSVFileParameter.cpp
@@ -30,6 +30,8 @@
 
 #include "ReadCSVFileParameter.hpp"
 
+#include "simplnx/Common/TypesUtility.hpp"
+
 namespace nx::core
 {
 // -----------------------------------------------------------------------------
@@ -83,4 +85,83 @@ Result<> ReadCSVFileParameter::validate(const std::any& value) const
   [[maybe_unused]] auto data = std::any_cast<ReadCSVData>(value);
   return {};
 }
+
+namespace SIMPLConversion
+{
+namespace
+{
+constexpr StringLiteral k_InputFilePathKey = "Wizard_InputFilePath";
+constexpr StringLiteral k_DataHeadersKey = "Wizard_DataHeaders";
+constexpr StringLiteral k_BeginIndexKey = "Wizard_BeginIndex";
+constexpr StringLiteral k_NumberOfLinesKey = "Wizard_NumberOfLines";
+constexpr StringLiteral k_DataTypesKey = "Wizard_DataTypes";
+constexpr StringLiteral k_DelimitersKey = "Wizard_Delimiters";
+constexpr StringLiteral k_HeaderLineKey = "Wizard_HeaderLine";
+constexpr StringLiteral k_HeaderIsCustomKey = "Wizard_HeaderIsCustom";
+constexpr StringLiteral k_HeaderUsesDefaultsKey = "Wizard_HeaderUseDefaults";
+constexpr StringLiteral k_ConsecutiveDelimitersKey = "Wizard_ConsecutiveDelimiters";
+
+std::vector<DataType> ConvertDataTypeStrings(const std::vector<std::string>& dataTypes)
+{
+  std::vector<DataType> output;
+
+  for(usize i = 0; i < dataTypes.size(); i++)
+  {
+    try
+    {
+      output.push_back(nx::core::StringToDataType(dataTypes[i]));
+    } catch(const std::exception& e)
+    {
+    }
+  }
+
+  return output;
+}
+
+std::vector<char> ConvertToChars(const std::string& string)
+{
+  return std::vector<char>(string.begin(), string.end());
+}
+} // namespace
+
+Result<ReadASCIIWizardDataFilterParameterConverter::ValueType> ReadASCIIWizardDataFilterParameterConverter::convert(const nlohmann::json& json)
+{
+  std::vector<std::string> dataTypeStrings = json[k_DataTypesKey].get<std::vector<std::string>>();
+  for(auto& dataTypeStr : dataTypeStrings)
+  {
+    if(dataTypeStr == "double")
+    {
+      dataTypeStr = "float64";
+    }
+    else if(dataTypeStr == "float")
+    {
+      dataTypeStr = "float32";
+    }
+  }
+
+  ValueType value;
+  value.inputFilePath = json[k_InputFilePathKey].get<std::string>();
+  value.customHeaders = json[k_DataHeadersKey].get<std::vector<std::string>>();
+  value.startImportRow = json[k_BeginIndexKey].get<int32>();
+  value.dataTypes = ConvertDataTypeStrings(dataTypeStrings);
+  value.delimiters = ConvertToChars(json[k_DelimitersKey].get<std::string>());
+  value.headersLine = json[k_HeaderLineKey].get<int32>();
+
+  bool headerIsCustom = json[k_HeaderIsCustomKey].get<bool>();
+  bool headerUsesDefaults = json[k_HeaderUsesDefaultsKey].get<bool>();
+
+  if(headerIsCustom)
+  {
+    value.headerMode = ValueType::HeaderMode::CUSTOM;
+  }
+  else
+  {
+    value.headerMode = ValueType::HeaderMode::LINE;
+  }
+
+  value.consecutiveDelimiters = static_cast<bool>(json[k_ConsecutiveDelimitersKey].get<int32>());
+
+  return {std::move(value)};
+}
+} // namespace SIMPLConversion
 } // namespace nx::core
